requirements.txt
Flask
python-dotenv
google-generativeai
Pillow
gunicorn
Flask-Cors 